requirements.txt
attrs==25.3.0
boto3==1.37.37
botocore==1.37.37
Cerberus==1.3.7
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lark==1.2.2
numpy==2.2.4
pandas==2.2.3
prettytable==3.16.0
python-dateutil==2.9.0.post0
python-hcl2==7.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
s3transfer==0.11.5
six==1.17.0
tabulate==0.9.0
tomli==2.2.1
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
